requirements.txt
python-telegram-bot==21.9
httpx==0.28.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.7.1
structlog==24.4.0
